+++ CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixed
-- Search Progress bar and Count Per Page components in "search-result"
+### Added
+- Be able to use `search-products-progress-bar` in `search-result`
 
 ## [3.62.2] - 2020-06-23
 ### Fixed